move classes for masks and masked circuits
